Make visual commands work over tramp
Currently if you use `eat-eshell-mode` and try to run a visual
command like `top` on a remote host, everything works as
expected. However, if you try to enable
`eat-eshell-visual-command-mode` and do the same thing, the
execution will fail. If you look at the error message, it
contains the full path to the binary including the TRAMP
prefix (e.g. `/ssh:<host>:/path/to/binary`). Of course, passing
that binary path to `make-process` to execute on a remote host will
fail.

The reason why this works fine for `eat-eshell-mode` is because
it makes an advice around `eshell-gather-process-output` and hijacks
the `make-process` call. The `eat-eshell-visual-command-mode`
on the other hand just replaces `eshell-exec-visual`, which
doesn't seem to be TRAMP-aware by default.

This explains why you need to do `(setq eshell-visual-commands nil)`
in order to make certain tools like `top` run even under
`eat-eshell-mode`.

The fix in this patch is simple and (hopefully) non-intrusive. I
just check that the binary we're trying to execute in
`eat--eshell-exec-visual` is on a TRAMP path and then I just
take the "local" name without the prefix. Fortunately,
`make-process` is aware of TRAMP at the level of
`current-directory` and will run the binary on the remote host
if `current-directory` is remote.

Otherwise the fix should be relatively safe, as it only works
for TRAMP paths. This can probably be fixed in eshell itself as
well, but since Eat copies the function, it will still need
the same fix.