Map OSX native paste command to use eat-yank
First off, Thank you for this wonderfull emacs mode. I have been using
it for over a year and I really like it for its speed and eshell
integration.

I pushed this PR because it took me sometime to figure out an issue
where I used my voice to instruct claude code (using claude-code.el
mode that uses eat) to send prompt instruction. It failed with this
error message: ``` insert-for-yank: Text is read-only ``` The issue
was that the software is using system wide hotkey for OSX to pasting,
which is s-v (Command + v), which is calling `yank` fn instead of
`eat-yank` when I am in `eat-semi-char-mode`.

However, it is not an issue of eat since it can be just added using
following in the config:
```
(define-key eat-semi-char-mode-map [?\s-v] #'eat-yank)
```

I would leave it to you to either merge or close but this should be
documented somewhere so someone might benefit from this.