Handle TERMINFO settings on remote hosts
Currently eat does not work over tramp by default due to
missing TERMINFO. The work-arounds are to either manually copy
over the TERMINFO and change the path (which causes it to not
work on the local machine), or set `TERM` to `xterm` and loose
some of the features.

This change makes it so that eat will first check for the
TERMINFO on the remote host, and if it is not present it will
copy it over.

closes #119